config: add --use-cache command line option
A cache file is used when connecting to a server. If the server closes
before commiting all changes made by a client then the cache will
contain edits not on the server, but will continue to be shown on the
client, until the cache is updated or deleted.

Use this new option to check if that has happened:

    ./piworld --server 192.168.1.64 --use-cache 0