Fix an out-of-memory error
When moving around a world with a lot of edits a chunk may sometimes
appear to go missing. This is because the GPU has denied the allocation
of memory for that chunk (glBufferData in gen_buffer).

To reduce the likelihood of that happening I've reduced the chunk delete
radius by 1.