Begin PWPI: the PiWorld Programming Interface
PWPI allows other processes to query and edit the world using a basic
text protocol, it is based on the existing protocol used by server.py.

The libevent library is used so you may need to install its dev package
before rebuilding piworld:

    $ sudo apt-get install libevent-dev
    $ cmake .
    $ make

The interface will listen to localhost connections only, over port
11760.

An example using netcat, with piworld already running:

    $ nc localhost 11760
    Q,P
    5.051038,12.750000,0.570777,1.728321,-0.395784,8,11,1,2,3
    P,20,12,20
    B,24,12,20,4

The example above queries the state of the player (returning position,
view angle, the location and type of block in the crosshairs and item in
hand), sets the player position to (20, 12, 20) and finally sets the
block at (24, 12, 20) to 4 (currently brick).