Add ETC1 compressed version of sign texture
Using this ETC1 compressed texture reduces GPU memory usage for this one
texture from 260K to 36K.

To check this (on Pi 0-3, not available on Pi 4):

  $ sudo vcdbg reloc | grep Texture

there should be 2 texture entries at 260K and one now at 36K.

ETC1 compression works well for the black and white sign texture, but
reduces quality too much for the remaining colour textures, so the sign
texture will be the only texture to use ETC1.

Compressonator was used to create the ETC1 file (with a DDS header).

see: https://github.com/GPUOpen-Tools/Compressonator

To recreate the ETC1 file you'll need ImageMagick (to flip the image)
and Compressonator (to compress it):

  $ convert -flip textures/sign.png sign-flip.png
  $ CompressonatorCLI-bin -fd ETC_RGB sign-flip.png sign
  $ mv sign.dds bin/