Reduce GPU memory use of chunk data
Use half floats for the vector data in chunks instead of floats. This
uses nearly half the amount of GPU memory needed to store each chunk.

Only use half floats on Pi 0-3 with the non-Mesa driver as half floats
with the Pi Mesa driver do not work on the Pi 3 and on the Pi 4 are
noticeably slower than using full floats.

To test half floats with the Mesa driver pass `--hfloat 1` on the
command line.