Change fullscreen size to match Pi performance
When using the Mesa driver and in fullscreen mode a maximum width of
about 720 pixels appears to be the performance sweet spot for the Pi.

The xrandr command line tool is used to get and set the fullscreen size.

To test other fullscreen sizes use the new command line option:

  --fullscreen-size WxH

where WxH is a valid size as reported for your screen by xrandr.

Beware that some sizes may be misreported as supported, if your screen
becomes unusable use `xrandr -s 0` to return to the first supported
size.