Add menu system
You can now open a menu by pressing Escape on a keyboard, the Start
button on a gamepad or right click on a control block.

Basic main and options menus are available with a limited list of
options to toggle.

Split-screen players can control their own set of menus with whatever
input device is bound to that player (keyboard, mouse or gamepad).