pg: pi: brcm: attempt to correct es2 window placement
On some screen setups piworld was showing its 3D display either badly
aligned or not at all.

Fix a bug in which width and height are unsigned int's, but the result
for xy position is a signed int by casting width and height to an
unsigned int for xy calculations.